requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
Django==4.2.13
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipython==8.18.1
jedi==0.19.1
Markdown==3.6
markdown2==2.4.13
matplotlib-inline==0.1.7
multidict==6.0.5
openai==1.34.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
twilio==9.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.19.2
